# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/main/applications": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Last applications",
                "description": "Get two last applications",
                "operationId": "82ba9339322faa082194cec20817d38b",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Натюрморт"
                                        },
                                        "user": {
                                            "type": "string",
                                            "example": "Иван Иванов"
                                        },
                                        "user_avatar": {
                                            "type": "string",
                                            "example": "/storage/img/avatar.png"
                                        },
                                        "application_category": {
                                            "type": "string",
                                            "example": "Для всех"
                                        },
                                        "description": {
                                            "type": "text",
                                            "example": "Нарисовать натюрморт"
                                        },
                                        "payment": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "5000"
                                        },
                                        "writing_technique": {
                                            "type": "string",
                                            "example": "Акварель"
                                        },
                                        "deadline": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "20.04.2022"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "3 weeks ago"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/applications": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Applications",
                "description": "Get all applications",
                "operationId": "a0e3611298a621e7c3b63e35c5c8777f",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Натюрморт"
                                        },
                                        "user": {
                                            "type": "string",
                                            "example": "Иван Иванов"
                                        },
                                        "user_avatar": {
                                            "type": "string",
                                            "example": "/storage/img/avatar.png"
                                        },
                                        "application_category": {
                                            "type": "string",
                                            "example": "Для всех"
                                        },
                                        "description": {
                                            "type": "text",
                                            "example": "Нарисовать натюрморт"
                                        },
                                        "payment": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "5000"
                                        },
                                        "writing_technique": {
                                            "type": "string",
                                            "example": "Акварель"
                                        },
                                        "deadline": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "20.04.2022"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "3 weeks ago"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "customer"
                ],
                "summary": "Applications",
                "description": "Create application",
                "operationId": "9a9802f924b1d0ab3108b15b76375abb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "genre_id": {
                                        "type": "required, exists:genres,id"
                                    },
                                    "description": {
                                        "type": "required, string, min:10, max:300"
                                    },
                                    "payment": {
                                        "type": "required, numeric, min:1, max:999999"
                                    },
                                    "writing_technique": {
                                        "type": "required, string, max:50"
                                    },
                                    "deadline": {
                                        "type": "required, date"
                                    },
                                    "application_category_id": {
                                        "type": "required, exists:application_categories,id"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "reason": "1",
                                    "description": "Нарисовать натюрморт",
                                    "payment": "1000",
                                    "writing_technique": "Масло",
                                    "deadline": "10.10.2022",
                                    "application_category_id": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Заявка успешно создана"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}